Move QueryInfo class into its own module.
It may seem like we're back to where we started with the separate
QueryStore class, but what we've actually done is consolidated a lot of
important logic into the existing QueryInfo abstraction.
